chore: gen data serverside
